Separate Guitar class into its own file
